Change naming to respect Rspec ideology
